tools/agentic-ai/requirements.txt

pyyaml==6.0.2
requests==2.32.3
lxml==5.2.2
colorama==0.4.6